requirements.txt
asgiref==3.3.1
astroid==2.4.2
boto3==1.17.77
botocore==1.20.77
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.4
django-bootstrap-form==3.4
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-storages==1.11.1
Django-Verify-Email==1.0.7
gunicorn==20.1.0
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
psycopg2==2.8.6
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.5
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
